requirements.txt
amqp==2.5.2
aniso8601==4.1.0
appdirs==1.4.3
argh==0.26.2
attrs==19.2.0
beautifulsoup4==4.8.1
billiard==3.6.1.0
black==19.3b0
bs4==0.0.1
cachetools==3.1.1
celery==4.3.0
celerybeat-mongo
certifi==2019.9.11
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
colorama==0.4.1
decorator==4.4.0
diario==0.2.0
dnspython==2.0.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
future==0.17.1
google==2.0.3
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
gunicorn==19.9.0
html5lib==1.0b10
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
instaloader==4.4.4
ipwhois==1.1.0
Jinja2==2.10
kombu==4.6.6
lxml==4.4.1
MarkupSafe==1.1.0
mongoengine
more-itertools==7.2.0
passlib==1.7.1
pathtools==0.1.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pymongo==3.9.0
PySocks==1.7.1
python-jose==3.2.0
PyJWT==1.7.1
python-whois==0.7.2
pytz==2018.9
PyYAML==5.2
redis==3.2.0
requests==2.22.0
requests-futures==1.0.0
requests-toolbelt==0.9.1
rope==0.14.0
rsa==4.0
sdklib==1.10.3
shodan==1.19.0
six==1.12.0
soupsieve==1.9.5
stem>=1.8.0
tld==0.12.2
threatconnect
toml==0.10.0
torrequest==0.1.0
tweepy==3.8.0
uritemplate==3.0.0
urllib3==1.25.7
validators==0.14.0
vine==1.3.0
watchdog==0.9.0
webencodings==0.5.1
Werkzeug==0.15.3
XlsxWriter==1.2.5
zipp==0.6.0
zxcvbn==4.4.28